requirements.txt
requests==2.31.0
requests-oauthlib==1.3.1
Pillow>=11.0.0
